+++ pudl_scrapers/spiders/ferc714.py
@@ -4,8 +4,8 @@ from pathlib import Path
 import scrapy
 from scrapy.http import Request
 
-from pudl import items
-from pudl.helpers import new_output_dir
+from pudl_scrapers import items
+from pudl_scrapers.helpers import new_output_dir
 
 
 class Ferc714Spider(scrapy.Spider):
