+++ scrapy.cfg
@@ -4,8 +4,8 @@
 # https://scrapyd.readthedocs.io/en/latest/deploy.html
 
 [settings]
-default = pudl.settings
+default = pudl_scrapers.settings
 
 [deploy]
 #url = http://localhost:6800/
-project = pudl
+project = pudl_scrapers
